refactor: move sut ID parsing out of DynamicSUTMaker and into SUTMetadata
